backend/pyproject.toml
[project]
name = "codeforge"
version = "0.1.0"
description = "AI Code Agent with Self-Repair"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "docker>=7.0.0",
    "redis>=5.0.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: unit tests",
    "integration: requires Docker/DB",
    "e2e: full stack tests",
]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
